Prompt name refactor
